requirements.txt
chromadb==0.5.3
openai==1.34.0
python-dotenv==0.19.2
sentence-transformers==3.0.1
streamlit==1.35.0